Working build system for kernel project
