+++ config/default_target.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+echo i686-elf
