+++ src/base/main.c
@@ -1,10 +1,7 @@
 #include <screen.h>
-#include <stddef.h>
-#include <stdint.h>
 #include <system.h>
+#include <string.h>
 
-
-/*TODO: implement kernel textmode from tutorial_kernel.c here as well! */
 /* This is a very simple main() function. All it does is sit in an
 *  infinite loop. This will be like our 'idle' loop */
 void kernel_main()
