+++ src/base/src/screen.c
@@ -1,3 +1,7 @@
+#include <stdbool.h>
+#include <stddef.h>
+#include <stdint.h>
+#include <string.h>
 #include <system.h>
 
 /* Hardware text mode color constants. */
