--- src/grub.cfg
@@ -1,3 +1,0 @@
-menuentry "myos" {
-	multiboot /boot/myos.bin
-}
