+++ src/lib/include/stdio.h
@@ -1,0 +1,8 @@
+#ifndef __STDIO_H
+#define __STDIO_H
+
+/*int printf(const char* __restrict, ...);
+int putchar(int);
+int puts(const char*);*/
+
+#endif
